+++ Sources/StickyEncoding/StorageContainerReaderWriter.swift
@@ -60,12 +60,10 @@ private enum Element {
     /// Element Header structure (will be written at the beginning of each Element)
     ///
     struct Header {
-        init(_ type: ContainerType, byteCount: Int) {
-            self.type  = type
-            self.byteCount = Int32(byteCount)
+        init(_ type: ContainerType) {
+            self.type     = type
         }
         let type: ContainerType
-        let byteCount: Int32
     }
 }
 
@@ -76,18 +74,32 @@ internal class StorageContainerWriter {
 
     // MARK: - Root level methods
 
-    @discardableResult
     @inline(__always)
-    static func write(_ storageContainer: StorageContainer, to buffer: UnsafeMutableRawBufferPointer) -> Int {
-        return write(storageContainer: storageContainer, to: buffer)
-    }
+    static func convert(_ storageContainer: StorageContainer) -> [UInt8] {
+        var bytes = Array<UInt8>()
 
-    @inline(__always)
-    static func byteCount(_ storageContainer: StorageContainer) -> Int {
-        return byteCount(storageContainer: storageContainer)
+        self.write(storageContainer: storageContainer, to: &bytes)
+
+        return bytes
     }
 
     // MARK: - Private methods
+
+    private static func pad<T>(_ buffer: inout [UInt8], for type: T.Type) {
+        let alignedOffset = align(offset: buffer.count, to: type)
+
+        if alignedOffset > buffer.count {
+            buffer.append(contentsOf: Array<UInt8>(repeating: 0x0, count: alignedOffset - buffer.count))
+        }
+    }
+
+    private static func write(header type: Element.ContainerType, to buffer: inout [UInt8]) {
+
+        /// Align Header if required
+        pad(&buffer, for: Element.Header.self)
+
+        withUnsafeBytes(of: Element.Header(type)) {  buffer.append(contentsOf: $0) }
+    }
 
     /// Write a container as an Element (Header + Container).
     ///
@@ -98,47 +110,18 @@ internal class StorageContainerWriter {
     /// ```
     ///
     /* Do not inline, recursively called */
-    private static func write(storageContainer: StorageContainer, to buffer: UnsafeMutableRawBufferPointer) -> Int {
-
-        /// Create a buffer that starts past the header which will be written last.
-        ///
-        let elementOffset = align(offset: MemoryLayout<Element.Header>.stride, to: UnsafeRawPointer.self)
-        let elementBuffer = UnsafeMutableRawBufferPointer(rebasing: buffer[elementOffset...])
-
-        /// The Element Header
-        var header: Element.Header
+    private static func write(storageContainer: StorageContainer, to buffer: inout [UInt8]) {
 
         /// Write -> Data
         switch storageContainer {
 
-        case let container as SingleValueContainer:    header = Element.Header(.singleValue,   byteCount: write(singleValue: container, to: elementBuffer)); break
-        case let container as UnkeyedStorageContainer: header = Element.Header(.unkeyed,       byteCount: write(unkeyed: container,     to: elementBuffer)); break
-        case let container as KeyedStorageContainer:   header = Element.Header(.keyed,         byteCount: write(keyed: container,       to: elementBuffer)); break
+        case let container as SingleValueContainer:    write(singleValue: container, to: &buffer); break
+        case let container as UnkeyedStorageContainer: write(unkeyed:     container, to: &buffer); break
+        case let container as KeyedStorageContainer:   write(keyed:       container, to: &buffer); break
         case is NullStorageContainer: fallthrough
-        default: header = Element.Header(.null, byteCount: write(null: NullStorageContainer.null, to: elementBuffer)); break
+        default:
+            write(null: NullStorageContainer.null, to: &buffer); break
         }
-
-        /// Write -> Type/byteCount
-        buffer.storeBytes(of: header, as: Element.Header.self)
-
-        return elementOffset + Int(header.byteCount)
-    }
-
-    /* Do not inline, recursively called */
-    private static func byteCount(storageContainer: StorageContainer) -> Int {
-
-        var bytes = align(offset: MemoryLayout<Element.Header>.stride, to: UnsafeRawPointer.self)
-
-        /// Write -> Data
-        switch storageContainer {
-
-        case let container as KeyedStorageContainer:   bytes += byteCount(keyed:   container); break
-        case let container as UnkeyedStorageContainer: bytes += byteCount(unkeyed: container); break
-        case let container as SingleValueContainer:    bytes += byteCount(singleValue:   container); break
-        case is NullStorageContainer: fallthrough
-        default: bytes += byteCount(null: NullStorageContainer.null); break
-        }
-        return bytes
     }
 
     // MARK: - `SingleValueContainer` methods.
@@ -156,14 +139,12 @@ internal class StorageContainerWriter {
     ///       Int32 | In32 | Determined by size
     /// ```
     @inline(__always)
-    private static func write(singleValue storageContainer: SingleValueContainer, to buffer: UnsafeMutableRawBufferPointer) -> Int {
-        storageContainer.write(to: buffer)
-        return storageContainer.byteCount
-    }
+    private static func write(singleValue storageContainer: SingleValueContainer, to buffer: inout [UInt8]) {
 
-    @inline(__always)
-    private static func byteCount(singleValue storageContainer: SingleValueContainer) -> Int {
-        return storageContainer.byteCount
+        /// Write the header
+        write(header: .singleValue, to: &buffer)
+
+        buffer.append(contentsOf: storageContainer.bytes)
     }
 
     // MARK: - `UnkeyedStorageContainer` methods.
@@ -179,29 +160,19 @@ internal class StorageContainerWriter {
     /// ```
     ///
     @inline(__always)
-    private static func write(unkeyed storageContainer: UnkeyedStorageContainer, to buffer: UnsafeMutableRawBufferPointer) -> Int {
+    private static func write(unkeyed storageContainer: UnkeyedStorageContainer, to buffer: inout [UInt8]) {
+
+        /// Write the header
+        write(header: .unkeyed, to: &buffer)
 
         /// Write -> Count
-        buffer.storeBytes(of: Int32(storageContainer.count), as: Int32.self)
-        var byteCount = align(offset: MemoryLayout<Int32>.stride, to: UnsafeRawPointer.self)
+        pad(&buffer, for: Int32.self)
+        buffer.append(contentsOf: withUnsafeBytes(of: Int32(storageContainer.count)) { Array($0) })
 
         for index in 0..<storageContainer.count {
-
             /// Write -> Element n
-            let elementByteCount =  write(storageContainer: storageContainer[index], to: UnsafeMutableRawBufferPointer(rebasing: buffer[byteCount...]))
-            byteCount = align(offset: byteCount + elementByteCount, to: UnsafeRawPointer.self)
+            write(storageContainer: storageContainer[index], to: &buffer)
         }
-        return byteCount
-    }
-
-    @inline(__always)
-    private static func byteCount(unkeyed storageContainer: UnkeyedStorageContainer) -> Int {
-        var bytes = align(offset: MemoryLayout<Int32>.stride, to: UnsafeRawPointer.self)
-
-        for index in 0..<storageContainer.count {
-            bytes = align(offset: bytes + byteCount(storageContainer: storageContainer[index]), to: UnsafeRawPointer.self)
-        }
-        return bytes
     }
 
     // MARK: - `KeyedStorageContainer` methods.
@@ -217,35 +188,23 @@ internal class StorageContainerWriter {
     /// ```
     ///
     @inline(__always)
-    private static func write(keyed storageContainer: KeyedStorageContainer, to buffer: UnsafeMutableRawBufferPointer) -> Int {
+    private static func write(keyed storageContainer: KeyedStorageContainer, to buffer: inout [UInt8]) {
+
+        /// Write the header
+        write(header: .keyed, to: &buffer)
 
         /// Write -> count
-        buffer.storeBytes(of: Int32(storageContainer.count), as: Int32.self)
-        var offset = align(offset: MemoryLayout<Int32>.stride, to: UnsafeRawPointer.self)
+        pad(&buffer, for: Int32.self)
+        buffer.append(contentsOf: withUnsafeBytes(of: Int32(storageContainer.count)) { Array($0) })
 
         for (key, value) in storageContainer {
 
             /// Write - Key n
-            let keyByteCount = write(string: key, to: UnsafeMutableRawBufferPointer(rebasing: buffer[offset...]))
-            offset = align(offset: offset + keyByteCount, to: UnsafeRawPointer.self)
+            write(string: key, to: &buffer)
 
             /// Write -> Value n
-            let valueByteCount = write(storageContainer: value, to: UnsafeMutableRawBufferPointer(rebasing: buffer[offset...]))
-            offset = align(offset: offset + valueByteCount, to: UnsafeRawPointer.self)
+            write(storageContainer: value, to: &buffer)
         }
-        return offset
-    }
-
-    @inline(__always)
-    private static func byteCount(keyed storageContainer: KeyedStorageContainer) -> Int {
-
-        var bytes = align(offset: MemoryLayout<Int32>.stride, to: UnsafeRawPointer.self)
-
-        for (key, value) in storageContainer {
-            bytes = align(offset: bytes + byteCount(string: key), to: UnsafeRawPointer.self)
-            bytes = align(offset: bytes + byteCount(storageContainer: value),  to: UnsafeRawPointer.self)
-        }
-        return bytes
     }
 
     // MARK: - `String` methods.
@@ -254,38 +213,27 @@ internal class StorageContainerWriter {
     /// Writes a string (used for dictionary keys.)
     ///
     @inline(__always)
-    private static func write(string: String, to buffer: UnsafeMutableRawBufferPointer) -> Int {
+    private static func write(string: String, to buffer: inout [UInt8]) {
         let utf8 = string.utf8
 
         /// Store the size of the string
-        buffer.storeBytes(of: Int32(utf8.count), as: Int32.self)
+        pad(&buffer, for: Int32.self)
+        buffer.append(contentsOf: withUnsafeBytes(of: Int32(utf8.count)) { Array($0) })
 
         /// Store the value
-        var offset = align(offset: MemoryLayout<Int32>.stride, to: MemoryLayout<Unicode.UTF8.CodeUnit>.self)
+        pad(&buffer, for: Unicode.UTF8.CodeUnit.self)
+
         for codeUnit in utf8 {
-            buffer.storeBytes(of: codeUnit, toByteOffset: offset, as: Unicode.UTF8.CodeUnit.self)
-            offset += MemoryLayout<Unicode.UTF8.CodeUnit>.stride
+            buffer.append(contentsOf: withUnsafeBytes(of: codeUnit) { Array($0) })
         }
-        return offset
-    }
-
-    @inline(__always)
-    private static func byteCount(string: String) -> Int {
-        let bytes = align(offset: MemoryLayout<Int32>.stride, to: MemoryLayout<Unicode.UTF8.CodeUnit>.self)
-
-        return bytes + (MemoryLayout<Unicode.UTF8.CodeUnit>.stride * string.count)
     }
 
     // MARK: - `NullStorageContainer` methods.
 
     @inline(__always)
-    private static func write(null: NullStorageContainer, to buffer: UnsafeMutableRawBufferPointer) -> Int {
-        return byteCount(null: null)
-    }
-
-    @inline(__always)
-    private static func byteCount(null: NullStorageContainer) -> Int {
-        return 0    /// Nothing to store for null
+    private static func write(null: NullStorageContainer, to buffer: inout [UInt8]) {
+        /// Write the header which is all we need to do
+        write(header: .null, to: &buffer)
     }
 }
 
@@ -302,11 +250,8 @@ internal class StorageContainerReader {
     /// - Returns: `StorageContainer` representation of the raw buffer.
     ///
     @inline(__always)
-    static func read(from buffer: UnsafeRawBufferPointer) throws -> StorageContainer {
-
-        let (container, _) = try read(from: buffer)
-
-        return container
+    static func convert(_ bytes: [UInt8]) throws -> StorageContainer {
+        return try bytes.withUnsafeBytes({ try read(from: $0[0...]).0 })
     }
 
     // MARK: - Private implementation
@@ -322,34 +267,31 @@ internal class StorageContainerReader {
     /// ```
     ///
     /* Do not inline, recursively called */
-    private static func read(from buffer: UnsafeRawBufferPointer) throws -> (StorageContainer, Int) {
+    private static func read(from buffer: Slice<UnsafeRawBufferPointer>) throws -> (StorageContainer, Int)   {
+
+        var offset = align(offset: buffer.startIndex, to: Element.Header.self)
 
         /// Ensure we can load the header
-        guard buffer.count >= MemoryLayout<Element.Header>.stride
+        guard buffer.endIndex > offset && buffer.endIndex - offset >= MemoryLayout<Element.Header>.size
             else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Binary data does not contian a proper header.")) }
 
         /// Load the buffer header
-        let header = buffer.load(as: Element.Header.self)
+        let header = UnsafeRawBufferPointer(rebasing: buffer[offset...]).load(as: Element.Header.self)
 
-        /// Ensure the buffer is large enough based on the header byte count
-        guard buffer.count >= MemoryLayout<Element.Header>.stride + Int(header.byteCount)
-            else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Binary data truncated or missing."))  }
+        offset += MemoryLayout<Element.Header>.size
 
-        let containerOffset = align(offset: MemoryLayout<Element.Header>.stride, to: UnsafeRawPointer.self)
-        let containerBuffer = UnsafeRawBufferPointer(rebasing: buffer[containerOffset..<containerOffset + Int(header.byteCount)])
-
-        var storageContainer: StorageContainer
+        let container: (StorageContainer, Int)
 
         switch header.type {
 
-        case .singleValue: storageContainer = try read(containerBuffer, as: SingleValueContainer.self);    break
-        case .unkeyed:     storageContainer = try read(containerBuffer, as: UnkeyedStorageContainer.self); break
-        case .keyed:       storageContainer = try read(containerBuffer, as: KeyedStorageContainer.self);   break
-        case .null:        storageContainer = try read(containerBuffer, as: NullStorageContainer.self);    break
+        case .singleValue: container = try read(singleValueContainer:  buffer[offset...]); break
+        case .unkeyed:     container = try read(unkeyedContainer:      buffer[offset...]); break
+        case .keyed:       container = try read(keyedContainer:        buffer[offset...]); break
+        case .null:        container = try read(nullContainer:         buffer[offset...]); break
         default:
             throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Binary data contains an unknown container type."))
         }
-        return (storageContainer, containerOffset + containerBuffer.count)
+        return (container.0, container.1 + (offset - buffer.startIndex))
     }
 
     ///
@@ -366,12 +308,11 @@ internal class StorageContainerReader {
     ///       Int32 | In32 | Determined by size
     /// ```
     @inline(__always)
-    private static func read(_ buffer: UnsafeRawBufferPointer, as: SingleValueContainer.Type) throws -> SingleValueContainer {
+    private static func read(singleValueContainer buffer: Slice<UnsafeRawBufferPointer>) throws -> (StorageContainer, Int)   {
 
-        guard buffer.count >= SingleValueContainer.HeaderSize
-            else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Single value container header missing or corrupt.")) }
+        let container = try SingleValueContainer(from: buffer)
 
-        return try SingleValueContainer(from: buffer)
+        return (container, container.byteCount)
     }
 
     ///
@@ -386,33 +327,33 @@ internal class StorageContainerReader {
     /// ```
     ///
     @inline(__always)
-    private static func read(_ buffer: UnsafeRawBufferPointer, as: UnkeyedStorageContainer.Type) throws -> UnkeyedStorageContainer {
+    private static func read(unkeyedContainer buffer: Slice<UnsafeRawBufferPointer>) throws -> (StorageContainer, Int) {
+
+        /// Find the count offset.
+        var offset = align(offset: buffer.startIndex, to: Int32.self)
 
         /// Ensure we can load the header
-        guard buffer.count >= MemoryLayout<Int32>.stride
+        guard buffer.endIndex > offset && buffer.endIndex - offset >= MemoryLayout<Int32>.size
             else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Unkeyed container header missing or corrupt.")) }
+
+        /// Read -> count
+        let elementCount = UnsafeRawBufferPointer(rebasing: buffer[offset...]).load(as: Int32.self)
 
         let container = UnkeyedStorageContainer()
 
-        /// Read -> count
-        let elementCount = Int32(buffer.load(as: Int32.self))
-        var elementOffset = align(offset: MemoryLayout<Int32>.stride, to: UnsafeRawPointer.self)
+        offset += MemoryLayout<Int32>.size
 
-        for n in 0..<elementCount {
+        for _ in 0..<elementCount {
 
             /// Read -> Element n
-
-            /// Ensure we can load the container
-            guard buffer.count >= elementOffset
-                else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Unkeyed container corrupt, expected \(elementCount) elements but found \(n - 1).")) }
-
-            let (element, byteCount) = try read(from: UnsafeRawBufferPointer(rebasing: buffer[elementOffset...]))
-            elementOffset = align(offset: elementOffset + byteCount, to: UnsafeRawPointer.self)
+            let (element, byteCount) = try read(from: buffer[offset...])
 
             /// Push the new element on to the result container
             container.push(element)
+
+            offset += byteCount
         }
-        return container
+        return (container, offset - buffer.startIndex)
     }
 
     ///
@@ -426,77 +367,80 @@ internal class StorageContainerReader {
     /// ```
     ///
     @inline(__always)
-    private static func read(_ buffer: UnsafeRawBufferPointer, as: KeyedStorageContainer.Type) throws -> KeyedStorageContainer {
+    private static func read(keyedContainer buffer: Slice<UnsafeRawBufferPointer>) throws -> (StorageContainer, Int) {
+
+        /// Find the count offset.
+        var offset = align(offset: buffer.startIndex, to: Int32.self)
 
         /// Ensure we can load the header
-        guard buffer.count >= MemoryLayout<Int32>.stride
+        guard buffer.endIndex > offset && buffer.endIndex - offset >= MemoryLayout<Int32>.size
             else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Keyed container header missing or corrupt.")) }
 
         /// Read -> count
-        let elementCount  = buffer.load(as: Int32.self)
-        var elementOffset = align(offset: MemoryLayout<Int32>.stride, to: UnsafeRawPointer.self)
+        let elementCount = UnsafeRawBufferPointer(rebasing: buffer[offset...]).load(as: Int32.self)
+
+        offset += MemoryLayout<Int32>.size
 
         let container = KeyedStorageContainer()
 
         for _ in 0..<elementCount {
 
             /// Read -> Key n
+            let (key, keyByteCount) = try read(key: buffer[offset...])
 
-            /// Be sure you can read the first element
-            guard buffer.count >= elementOffset
-                else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Keyed container Key missing or corrupt.")) }
-
-            let (key, keyByteCount) = try read(UnsafeRawBufferPointer(rebasing: buffer[elementOffset...]), as: String.self)
-            elementOffset = align(offset: elementOffset + keyByteCount, to: UnsafeRawPointer.self)
+            offset += keyByteCount
 
             /// Read -> Value n
-
-            /// Be sure you can read the first element
-            guard buffer.count >= elementOffset
-                else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Keyed container value missing or corrupt.")) }
-
-            let (value, valueByteCount) = try read(from: UnsafeRawBufferPointer(rebasing: buffer[elementOffset...]))
-            elementOffset = align(offset: elementOffset + valueByteCount, to: UnsafeRawPointer.self)
+            let (value, valueByteCount) = try read(from: buffer[offset...])
 
             /// Assign the key/value
             container[key] = value
+
+            offset += valueByteCount
         }
-        return container
+        return (container, offset - buffer.startIndex)
     }
 
     ///
     ///  Reads a `String` from `buffer` (used for dictionary key storage).
     ///
     @inline(__always)
-    private static func read(_ buffer: UnsafeRawBufferPointer, as: String.Type) throws -> (String, Int) {
-        var utf8: [Unicode.UTF8.CodeUnit] = []
+    private static func read(key buffer: Slice<UnsafeRawBufferPointer>) throws -> (String, Int) {
+
+        /// Find the count offset.
+        var offset = align(offset: buffer.startIndex, to: Int32.self)
 
         /// Ensure we can load the header
-        guard buffer.count >= MemoryLayout<Int32>.stride
+        guard buffer.endIndex > offset && buffer.endIndex - offset >= MemoryLayout<Int32>.size
             else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Key value header missing or corrupt.")) }
 
-        let count = Int(buffer.load(as: Int32.self))
+        /// Read -> count
+        let count = UnsafeRawBufferPointer(rebasing: buffer[offset...]).load(as: Int32.self)
 
         guard count > 0
             else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Key value header missing or corrupt.")) }
 
+        offset = align(offset: offset + MemoryLayout<Int32>.size, to: Unicode.UTF8.CodeUnit.self)
+
         /// Ensure we can load the value
-        guard buffer.count >= count
+        guard buffer.endIndex > offset && buffer.endIndex - offset >= count
             else { throw DecodingError.dataCorrupted(DecodingError.Context(codingPath: [], debugDescription: "Key value corrupt, expected \(count) bytes but found \(buffer.count).")) }
 
-        let offset = align(offset: MemoryLayout<Int32>.stride, to: MemoryLayout<Unicode.UTF8.CodeUnit>.self)
+        var utf8: [Unicode.UTF8.CodeUnit] = []
 
-        for i in 0..<count {
-            utf8.append(buffer.load(fromByteOffset: offset + i, as: Unicode.UTF8.CodeUnit.self))
+        for _ in 0..<count {
+            utf8.append(UnsafeRawBufferPointer(rebasing: buffer[offset...]).load(as: Unicode.UTF8.CodeUnit.self))
+
+            offset += MemoryLayout<Unicode.UTF8.CodeUnit>.stride
         }
-        return (String(decoding: utf8, as: UTF8.self), offset + (MemoryLayout<Unicode.UTF8.CodeUnit>.stride * count))
+        return (String(decoding: utf8, as: UTF8.self), offset - buffer.startIndex)
     }
 
     ///
     /// Reads a `NullStorageContainer` from `buffer`.
     ///
     @inline(__always)
-    private static func read(_ buffer: UnsafeRawBufferPointer, as: NullStorageContainer.Type) throws -> NullStorageContainer {
-        return NullStorageContainer.null
+    private static func read(nullContainer buffer: Slice<UnsafeRawBufferPointer>) throws -> (StorageContainer, Int) {
+        return (NullStorageContainer.null, 0)
     }
 }